requirements.txt
flask
sqlalchemy
flask-sqlalchemy
pymysql
flask-migrate
pyjwt